Add user authorisation for altering stock

+++ stock/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -7,8 +8,13 @@ from .models import Product, Category
 from .forms import ProductForm
 
 
+@login_required
 def add_product(request):
     ''' Add a product to the store '''
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You are not authorised to do this.')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -28,15 +34,22 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     ''' Edit a product '''
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You are not authorised to do this.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
             form.save()
             messages.success(request, 'Updated product successfully')
-            return redirect(reverse('product_detail', args=[product.id]))
+            return redirect(reverse('home'))
         else:
             messages.error(request, 'Failed to update product. Please try again.')
     else:
@@ -52,9 +65,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     ''' Deletes the selected stock item '''
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You are not authorised to do this.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Stock has been deleted.')
-    return redirect(reverse('products'))
+    return redirect(reverse('home'))
